refactor(edge): split control api runtime modules
